pyproject.toml
[project]
name = "dashpint"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests",
    "pyyaml"
]
